chore(dashboard): unify tooltips
